some addition on one device and muti
