+++ module/device/connection.py
@@ -46,12 +46,14 @@ def retry(func):
                 logger.error(e)
 
                 def init():
+                    self.adb_client.reboot()
                     self.adb_disconnect(self.serial)
                     self.adb_connect(self.serial)
             # AdbError
             except AdbError as e:
                 if handle_adb_error(e):
                     def init():
+                        self.adb_client.reboot()
                         self.adb_disconnect(self.serial)
                         self.adb_connect(self.serial)
                 else:
@@ -68,7 +70,6 @@ def retry(func):
 
                 def init():
                     pass
-
         logger.critical(f'Retry {func.__name__}() failed')
         raise RequestHumanTakeover
 
@@ -467,6 +468,78 @@ class Connection:
         del_cached_property(self, 'hermit_session')
         del_cached_property(self, 'minitouch_builder')
         del_cached_property(self, 'reverse_server')
+
+    def adb_device_num(self):
+        """
+           count current device number
+        """
+        with self.adb_client._connect() as c:
+            count = 0
+            c.send_command("host:devices")
+            c.check_okay()
+            output = c.read_string_block()
+            for line in output.splitlines():
+                parts = line.strip().split("\t")
+                if len(parts) != 2:
+                    continue
+                if parts[1] == 'device':
+                    count = count + 1
+            return count
+
+    def adb_reboot(self):
+        """
+           Reboot adb client
+        """
+        if self.adb_device_num() == 0:
+            # Kill current client
+            self.adb_client.server_kill()
+            # Init adb client
+            self.adb_client = AdbClient('127.0.0.1', 5037)
+            # Monkey patch to custom adb
+            adbutils.adb_path = lambda: self.adb_binary
+            # Remove global proxies, or uiautomator2 will go through it
+            for k in list(os.environ.keys()):
+                if k.lower().endswith('_proxy'):
+                    del os.environ[k]
+            self.adb_client = AdbClient('127.0.0.1', 5037)
+
+            # Parse custom serial
+            self.serial = str(self.config.Emulator_Serial)
+            if "bluestacks4-hyperv" in self.serial:
+                self.serial = self.find_bluestacks4_hyperv(self.serial)
+            if "bluestacks5-hyperv" in self.serial:
+                self.serial = self.find_bluestacks5_hyperv(self.serial)
+            if "127.0.0.1:58526" in self.serial:
+                logger.warning('Serial 127.0.0.1:58526 seems to be WSA, '
+                               'please use "wsa-0" or others instead')
+                raise RequestHumanTakeover
+            if "wsa" in self.serial:
+                self.serial = '127.0.0.1:58526'
+                if self.config.Emulator_ScreenshotMethod != 'uiautomator2' \
+                        or self.config.Emulator_ControlMethod != 'uiautomator2':
+                    with self.config.multi_set():
+                        self.config.Emulator_ScreenshotMethod = 'uiautomator2'
+                        self.config.Emulator_ControlMethod = 'uiautomator2'
+            devices = list(self.iter_device())
+            # Connect
+            self.adb_connect(self.serial)
+            if self.config.Emulator_Serial == 'auto':
+                if len(devices) == 0:
+                    raise RequestHumanTakeover
+                elif len(devices) == 1:
+                    self.serial = devices[0].serial
+                    del_cached_property(self, 'adb')
+                else:
+                    raise RequestHumanTakeover
+            # Package
+            self.package = self.config.Emulator_PackageName
+            if self.package == 'auto':
+                self.detect_package(set_config=False)
+            else:
+                set_server(self.package)
+        else:
+            self.adb_disconnect(self.serial)
+            self.adb_connect(self.serial)
 
     def install_uiautomator2(self):
         """
